+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.8] - 2019-02-18
+
+### Add
+- Show devivative mean value
+
+### Fix
+- Show derivative graph mean and center in derivative mod
+- Calculating of center on gravity center
+
+## [0.4.7] - 2019-02-15
+
+### Add
+- Show devivative gravity center UMP
+
 ## [0.4.6] - 2019-02-11
 
 ### Add